+++ package.json
@@ -2,12 +2,17 @@
   "name": "zeronet",
   "version": "0.0.1-alpha18",
   "description": "ZeroNet JS",
-  "main": "index.js",
+  "main": "src/index.js",
+  "bin": {
+    "zeronet": "zeronet.js"
+  },
   "directories": {
     "lib": "lib",
     "test": "test"
   },
-  "browser": "browser.js",
+  "browser": {
+    "./src/bundles/node.js": "./src/bundles/browser.js"
+  },
   "dependencies": {
     "colors": "^1.1.2",
     "libp2p-tcp": "^0.11.1",
@@ -66,9 +71,6 @@
     "js",
     "p2p"
   ],
-  "bin": {
-    "zeronet": "zeronet.js"
-  },
   "pkg": {
     "scripts": [
       "node_modules/zeronet-swarm/hacks/**"
