+++ src/misc/common.js
