+++ src/misc/hacky-logs.js
