+++ tools/debug/debug.sh
@@ -1,1 +1,1 @@
-RUNINCWD=1 DEBUG_PACKETS=1 nodemon -x "DEBUG=zeronet*,libp2p* node zeronet" -d 2
+RUNINCWD=1 DEBUG_PACKETS=1 nodemon -x "DEBUG=zeronet*,libp2p* node src/bin" -d 2
