+++ tools/docs-gen/bin/document.js
