+++ tools/docs-gen/bin/index.js
