+++ tools/docs-gen/lib/index.js
