+++ tools/docs-gen/lib/renderable.js
